requirements.txt
semantic-kernel==1.28.1
chainlit
googlemaps
python-dotenv
asyncio
semantic-kernel-plugins
ipinfo